mute [nfc]: Move isTopicMuted to new muteModel module.
This is really part of the interface of the `MuteState` type.
So this is a more appropriate home for it than a module for
"message utilities".
