+++ src/utils/message.js
@@ -3,9 +3,6 @@ import type { Narrow, Message, MuteState, Outbox, Subscription } from '../types'
 import { isHomeNarrow, isTopicNarrow, isMentionedNarrow } from './narrow';
 import { streamNameOfStreamMessage } from './recipient';
 
-export const isTopicMuted = (stream: string, topic: string, mute: MuteState = []): boolean =>
-  mute.some(x => x[0] === stream && x[1] === topic);
-
 export const shouldBeMuted = (
   message: Message | Outbox,
   narrow: Narrow,
